refactor: move multer configuration to a separate middleware file and clean up routes
