+++ .gitignore
@@ -1,7 +1,5 @@
 /node_modules
-
 .env
-
 .DS_Store
 .vscode
