+++ routes/users.js
@@ -1,5 +1,5 @@
 import express from "express";
-import multer from "multer";
+import upload from "../middleware/multer.js";
 import {
   createUser,
   getUsers,
@@ -9,17 +9,6 @@ import {
 } from "../controllers/users.js";
 
 
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, "public/uploads/");
-  },
-  filename: function (req, file, cb) {
-    const timestamp = Date.now();
-    const cleanName = file.originalname.replace(/\s+/g, '-').toLowerCase();
-    cb(null, `${timestamp}-${cleanName}`);
-  },
-});
-const upload = multer({ storage });
 
 const router = express.Router();
